requirements.txt
# Core dependencies
flask==2.3.3
flask-session==0.5.0
gunicorn==21.2.0
werkzeug==2.3.7
python-dotenv==1.0.0

# Google API dependencies
google-api-python-client==2.108.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-analytics-data==0.17.1
google-ads==22.1.0

# Data processing
pandas==2.1.1
numpy==1.26.0